backend/requirements.txt
Flask
pymongo
flask-cors
pyjwt
werkzeug
python-dotenv